sfs-51: Adds bank holiday API URL to ADR.

+++ finance-api/internal/service/add_pending_collection.go
@@ -2,6 +2,7 @@ package service
 
 import (
 	"context"
+
 	"github.com/jackc/pgx/v5/pgtype"
 	"github.com/ministryofjustice/opg-sirius-supervision-finance-hub/finance-api/internal/auth"
 	"github.com/ministryofjustice/opg-sirius-supervision-finance-hub/finance-api/internal/store"
@@ -10,13 +11,13 @@ import (
 
 func (s *Service) AddPendingCollection(ctx context.Context, clientId int32, data shared.PendingCollection) error {
 	var collectionDate pgtype.Date
-	
+
 	_ = collectionDate.Scan(data.CollectionDate.Time)
 
 	err := s.store.CreatePendingCollection(ctx, store.CreatePendingCollectionParams{
 		ClientID:       clientId,
 		CollectionDate: collectionDate,
-		Amount:         int32(data.Amount),
+		Amount:         data.Amount,
 		CreatedBy:      ctx.(auth.Context).User.ID,
 	})
